Modernist Journals Project *The Crisis* (1910-1922) Social Network Data

W. E. B. Du Bois founded *The Crisis* in 1910 as the official magazine of the NAACP. The magazine was, as the Modernist Journals Project describes, "arguably the most widely read and influential periodical about race and social injustice in U.S. history." *The Crisis* is, in fact, still being published today.

Usage

mjp_crisis_network

Format

Nodes: 96; unimodal* Edges: 273; weighted; undirected

This dataset is from Melanie Walsh. Prepared for R by Benjamin Smith

Source

GitHub, <https://github.com/melaniewalsh/sample-social-network-datasets>

Details

The issues included in the MJP and this network data span from 1910-1922, a time period during which the magazine covered "most every facet of life for blacks in America, devoting special issues to such topics as women's suffrage, education, children, labor, homes, vacations, and the war." The magazine also became "an important crucible" for the literary and arts movement known as the Harlem Renaissance.

Notable authors who published in *The Crisis* and are included in this dataset: W. E. B. Du Bois, Jessie Fauset, Charles Chestnutt, Georgia Douglas Johnson, Claude McKay, and more.

For more information, see the Modernist Journals Project on *The Crisis*.